// src/payload-types.ts
export interface Website {
  id: number;
  title: string;
  url: string;
  admins?:
    | {
        admin: string;
        id?: string | null;
      }[]
    | null;
  editors?:
    | {
        editor?: string | null;
        id?: string | null;
      }[]
    | null;
  guests?:
    | {
        guest?: string | null;
        id?: string | null;
      }[]
    | null;
  enableDefaultViewports?: boolean | null;
  viewports?:
    | {
        label?: string | null;
        x?: number | null;
        y?: number | null;
        orientation?: boolean | null;
        id?: string | null;
      }[]
    | null;
  type: {
    mediaType: 'Website' | 'Figma' | 'Image' | 'Video' | 'PDF';
    injectionType: 'Proxy' | 'Extension' | 'Manual';
  };
  versions?:
    | {
        'page-links'?:
          | {
              'page-link': string;
              roomId: string;
              thread?:
                | {
                    'thread-id'?: string | null;
                    viewport?: string | null;
                    'viewport-Orientation'?: boolean | null;
                    comments?:
                      | {
                          'comment-id'?: string | null;
                          author?: string | null;
                          authorEmail?: string | null;
                          date?: string | null;
                          edited?: boolean | null;
                          editedAt?: string | null;
                          deleted?: boolean | null;
                          commentType?: ('comment' | 'task') | null;
                          commentStatus?: ('completed' | 'active') | null;
                          ping?: {
                            x?: number | null;
                            y?: number | null;
                            height?: number | null;
                            width?: number | null;
                          };
                          'text-edited'?: boolean | null;
                          'text-edit'?: {
                            selector?: string | null;
                            oldText?: string | null;
                            newText?: string | null;
                          };
                          message?: string | null;
                          id?: string | null;
                        }[]
                      | null;
                    id?: string | null;
                  }[]
                | null;
              id?: string | null;
            }[]
          | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}